refactoring, adding exception classes, testing methods

+++ src/main/java/enarleini/finance/Client/UserService.java
@@ -1,7 +1,8 @@
 package enarleini.finance.Client;
 
+import enarleini.finance.Client.exception.UserNotFoundException;
+import enarleini.finance.Client.exception.InvalidUserException;
 import enarleini.finance.config.JWTService;
-import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -14,7 +15,6 @@ import org.springframework.web.server.ResponseStatusException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
 
 @Service
 public class UserService {
@@ -34,9 +34,9 @@ public class UserService {
         if (repository.findByUsername(user.getUsername()) != null) {
             throw new ResponseStatusException(HttpStatus.CONFLICT, "Username already exists");
         }
-      user.setPassword(encoder.encode(user.getPassword()));
-      user.setRole(Roles.USER);
-      return repository.save(user);
+        user.setPassword(encoder.encode(user.getPassword()));
+        user.setRole(Roles.USER);
+        return repository.save(user);
     }
 
     public Map<String, String> verify(Users user) {
@@ -57,39 +57,40 @@ public class UserService {
     }
 
     public Users findClientById(Long id) {
-        return repository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Client not found."));
+        return repository.findById(id).orElseThrow(() -> new UserNotFoundException("Client not found with id " + id));
     }
 
     public Users findClientByUsername(String username) {
-        return repository.findByUsername(username);
+        Users user = repository.findByUsername(username);
+        if (user == null) {
+            throw new UserNotFoundException("Client not found with username " + username);
+        }
+        return user;
     }
 
-
     public void deleteClient(Long id) {
+        if (!repository.existsById(id)) {
+            throw new UserNotFoundException("Client not found with id " + id);
+        }
         repository.deleteById(id);
     }
 
     public void changeClientPassword(String username, String currentPassword, String newPassword) {
         Users existingClient = findClientByUsername(username);
-        if (existingClient != null) {
-            if (encoder.matches(currentPassword, existingClient.getPassword())) {
-                existingClient.setPassword(encoder.encode(newPassword));
-                repository.save(existingClient);
-            } else {
-                throw new RuntimeException("Current password is incorrect ");
-            }
+        if (encoder.matches(currentPassword, existingClient.getPassword())) {
+            existingClient.setPassword(encoder.encode(newPassword));
+            repository.save(existingClient);
         } else {
-            throw new IllegalArgumentException("Client with username " + username + " not found");
+            throw new InvalidUserException("Current password is incorrect");
         }
     }
 
     public void assignRoleToUser(String username, Roles role) {
         Users user = repository.findByUsername(username);
-        if (user != null) {
-            user.setRole(role);
-            repository.save(user);
-        } else {
-            throw new IllegalArgumentException("User with username " + username + " not found");
+        if (user == null) {
+            throw new UserNotFoundException("User not found with username " + username);
         }
+        user.setRole(role);
+        repository.save(user);
     }
 }
